Address lag in reporting output variables
This change adopts the new data exchange api provided natively by
energyplus to read output variables. This change at least partially
addresses a lag issue related to updating outputs. Still, the output
variables will only be updated when the simulation crosses a zone
timestep. For some use cases this might be an issue that will have to be
further addressed.