Add UnitDefinitions to modelDescription.xml file
There is more to this than the title indicates. Included in this is also
a system for attaching units to I/O variables, and then also a system to
perform unit conversion. Unit conversion is required because we need to
talk to EnergyPlus in EnergyPlus units, but expose variables through and
API that uses Modelica units.

There is also a lot of refactoring in this commit to improve
organization. Also a list of supported output variables has been created
and if a given idf files contains output variables that are not in the
supported list, then those unsupported variables are removed from the
FMU simulation.

close lbl-srg/modelica-buildings#1680