Add a 200 year "Spawn" RunPeriod
This change means that it is *almost impossible to reach the end of the
simulation period. Instead, spawn client will need to explicitly stop
the simulation.

This change also removes any preexisting run periods, although a future
change might leave them and in turn create an api to select which
runperiod that fmu will simulate.

close lbl-srg/modelica-buildings#1678