Set temp of controlled zones during warmup
In spawn there is the concept of connected and unconnected zones.
Connected zones are those that are exposed through the FMU interface and
are able to have an HVAC system attached through the FMU. Unconnected zones
are not exposed through the FMU, although zone related output variables can
be requested, however the FMU does not establish the zone temperature
like in connected zones.

Previously, the mixed air temperature of even "connected" zones was computed
by EnergyPlus during warmup. Connected zones were treated like
unconnected zones, which means warmup iterations were performed until
the zone air temperature stablized.

With this change, the temperature of connected zones is set by the FMU,
and EnergyPlus warmup does not interfere with the temperature
established by the FMU.