Adjust Spawn create energyplus fmu output options
Update the existing option --output-path, which is now intended to take
a full path including filename and extension to the EnergyPlus FMU that
will be created. Previously this option was intended to be a path to a
directory.

A new option --output-dir has been added. The --output-dir and
--output-path options are mutually exclusive, they cannot be used at the
same time. As the name indicates, the --output-dir should be given a
directory path where the EnergyPlus FMU should be created.

Documentation for these options is built into the spawn cli, and
available via spawn --help.

For both options, intermediate paths will be created as needed.
Exceptions will be thrown on error, such as permission denied, and there
is a new top level exception handler to report the exception to the
client.