Change EP working dir to resourceURI given to fm2Instantiate
Previously used instanceName as the working subdirectory but that is a
GUID, and not user friendly.

Note that instantiatiing the same EP FMU twice will not work as a result
of this change because the two instantiations will share the same
working dir. That said instantiating the EP FMU twice is not yet
supported anyway due to EnergyPlus global data.

ref lbl-srg/modelica-buildings#2053