Fix logging that involves "%" character
The FMI standard defines a logging callback function, where the last
named argument is a format string, and the remaining variable number of
unnamed arguments are the values to be replaced in the format string.
This follows the convention of the standard fprint function.

Spawn was previously sending the entire log message as the format string
and then no additional unnamed parameters.

This change is to send "%s" as the format string and then the content of
the log message as a single unnamed parameter.

Without this change there is a crash in the log message handling here:
https://github.com/lbl-srg/modelica-buildings/blob/bbb7cbf059bd0a21d942fac1e70ef19ada49271d/Buildings/Resources/src/ThermalZones/EnergyPlus/C-Sources/BuildingInstantiate.c#L515

close lbl-srg/modelica-buildings#2198