Update EnergyPlus and add tests for multi year sim
Previously the particular version of EnergyPlus tied to Spawn would hang
upon reaching the end of the weather file. Now the simulation will wrap
around to the beginning of the weather file as expected. The changes to
Spawn consist of updates to account for EnergyPlus chnages related to
global state refactor. There are functional changes to the Spawn
software.

Tests have been added to demonstrate multi-year tests via the EnergyPlus
FMI.

ref lbl-srg/modelica-buildings#2314