Remove isRunning check for setValue
There is a function "isRunning" that checks if EnergyPlus is running and
throws if it is not. Many of the Spawn class APIs use isRunning,
including setValue. This change is to make it so that isRunning is not
called for setValue. EnergyPlus does not need to be running in order to
set a value so the check was uncalled for.

ref #12