Update surface I/O to find matching surfaces
Interzone surfaces come in pairs. When setting surface temperatures,
this change finds the matching surface and sets the appropriate face
temperature.

Similarly when reporting surface heat flow for the outside face of an
interzone surface, the algorithm will retrieve the matching surface to
report the heat flow on that surface.