Be more strategic about exchanging data with EnergyPlus
This change avoids unnessecary calls to ::exchange if the input has not
chnaged. This avoids extra calls the EnergyPlus surface heat balance
calc. The standard numeric limits epsilon is used to define the
threshold to detect an input change.

A separate commit will make a Spawn configuration option to control the
surface heat balance convergence criteria.

ref lbl-srg/modelica-buildings#2437