Update test assertion to reflect change in behavior
As a result of d03d65a11a5e1c51355397f88f63b8a285d59bb2, if you
terminate (call fmi terminate), a model that is not running, the API
will return fmi2Error. This is a change in behavior as previously the
API would return fmi2OK.

This commit is to update the test assertions to reflect the current
behavior. I think that this is more correct, although it has been an open
question.