Fix error when there are no Output:Variable objects
The "idf preparation" which removes unsupported output variables (in an
effort to prevent warnings) was accessing the array of output variables using
unchecked methods. Here is the fix.

Thank you @helarga for originally reporting this. I think the reason we didn't
reproduce this back when it was first reported, is because it requires
both an idf file that does not request output variables, AND a spawn
model that does not request output. Spawn will automatically inject output requests
that are made from the Modelica side. That is my theory anyway. I'm
pretty sure it is squashed now.

ref lbl-srg/modelica-buildings#2088
ref lbl-srg/modelica-buildings#2512