Add test for all of the EnergyPlus example files
This test attempts to start and stop a simulation for every example file
that is shipped with EnergyPlus. This is all of the idf files contained
in the EnergyPlus testfiles directory, except those with an underscore
prefix.

There are 27 idf files that do not simulate because they exit early due
to fatal errors. These 27 files are identified in
test_example_files.cpp, and there are comments consisting of the error
file message that explains why they fail.

The reasons for simulation failure are summarized as the following:

* Preprocessor tokens (spawn does not handle epmacros at this time)
* The use of the external window program and supporting .dat files
* the use of ground preproc
* Runaway temps. Especially in the hospital models. Probably expected
for high loads and no hvac.
* The use of schedule:file which needs external files
* component costs input objects that reference HVAC that is not
available in spawn
* no zones in model

These items may be good content to document in release notes.

As a result of
https://github.com/NREL/Spawn/commit/d03d65a11a5e1c51355397f88f63b8a285d59bb2
all of these tests, even failing tests will exit cleanly and there will
be an eplusout.err file that explains the issue.