Add primitive cache variable to set path to optimica
An oem tarball of the Optimica distribution is required by the build
system, and up until this point the path was hard coded in cmake rules.
The end goal is to place the tarball on a file server for seemless
integration by the build system (as done for JModelica), however care is
taken with Optimica because it is proprietary. At this point access to
the tarball will be controlled, and distributed to developers on an as
needed basis.

This commit allows a developer to configure their unique location of the
Optimica oem distribution.