Update zone humidity ratio calculation
There have been a couple of long standing issues related to Spawn's
zone air moisture calculation, which has manifested as several
EnergyPlus warnings. The warning has been a symptom of a more
fundamental issue that is resolved in this commit.

1. For zones connected to Modelica, there is a zone air humidity ratio
input exposed by the FMU, but this input was not connected to the
EnergyPlus state variables.

2. For zones *not* connected to Modelica, the zone air humidity ratio
needs to be updated internally, just as we do for the zone air
temperature of unconnected zones. This has not happening for zone air
humidity ratio.

This commit addresses both items 1 and 2, and as a result for the test
model
"Buildings.ThermalZones.EnergyPlus.Examples.SingleFamilyHouse.AirHeating",
the warnings related to zone air humidity are cleared up.

ref lbl-srg/modelica-buildings#2220