Update MBL
This update introduced a change in the MBL that adjusted the model
organization.

Previously Spawn's EnergyPlus models were located in
Buildings/ThermalZones/EnergyPlus, however now the EnergyPlus version is
postfixed such as Buildings/ThermalZones/EnergyPlus_9_6_0.

Paths in the Spawn project have been adjusted to account for this change.

The EnergyPlus version is pulled from the EnergyPlus CMake project and a
function is made available to report the current EnergyPlus version
without relying on any compiled EnergyPlus software or libraries. This
is used for example in the spawn test cases.