Only add default people when required
Prior to this, every Spawn zone received a default people input object,
which was to avoid errors when accessing the people related outputs for
zones without people (such as attics).

As of version 24.2, EnergyPlus will issue a warning if there are
duplicate people inputs with different PMV, so to avoid such a possible
error the default people input is only provided when there is not an
existing people input for a particular zone.

ref https://github.com/lbl-srg/modelica-buildings/issues/3911#issuecomment-2452415560